Adds javascript helper javascript file
I started this on the wrong branch, but I'll continue on
a new, more apt branch tomorrow.  Basically, though, I'm
going to keep general helper functions in a file called
helpers.js. The first function I wrote hides all the views
except the one you pass as a parameter, as a more efficient way
to manage this SPA